--- lib/matplotlib/_backports/__init__.py
@@ -1,3 +1,0 @@
-"""
-Backported features from dependencies.
-"""
