+++ lib/mpl_toolkits/mplot3d/art3d.py
@@ -14,7 +14,7 @@ from six.moves import zip
 
 from matplotlib import (
     artist, cbook, colors as mcolors, lines, text as mtext, path as mpath)
-from matplotlib._backports import numpy as _backports_np
+from matplotlib.cbook import _backports
 from matplotlib.collections import (
     Collection, LineCollection, PolyCollection, PatchCollection,
     PathCollection)
@@ -774,7 +774,7 @@ def iscolor(c):
 
 def get_colors(c, num):
     """Stretch the color argument to provide the required number num"""
-    return _backports_np.broadcast_to(
+    return _backports.broadcast_to(
         mcolors.to_rgba_array(c) if len(c) else [0, 0, 0, 0],
         (num, 4))
